Refactor: move cairo-lang-macro types to submodule (#1194)
commit-id:ddf4c6b4

---

**Stack**:
- #1200
- #1195
- #1194 ⬅


⚠️ *Part of a stack created by [spr](https://github.com/ejoffe/spr). Do
not merge manually using the UI - doing so may have unexpected results.*
